task 102 commit 3

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,8 +8,8 @@
 int main(void)
 {
 	int i = 0;
-	int j = 1;
-	int h = 2;
+	long j = 1;
+	long h = 2;
 
 	while (i < 50)
 	{
